+++ FQHE/src/Hamiltonian/ParticleOnLatticeGenericHamiltonian.cc
@@ -59,7 +59,7 @@ using std::ostream;
 // memory = maximum amount of memory that can be allocated for fast multiplication (negative if there is no limit)
 // precalculationFileName = option file name where precalculation can be read instead of reevaluting them
 // hoppingOnly = evaluate only energy of hopping terms, excluding local potentials
-ParticleOnLatticeGenericHamiltonian::ParticleOnLatticeGenericHamiltonian(ParticleOnLattice* particles, int nbrParticles, LatticePhases *latticeGeometry, int nbrFluxQuanta, double contactInteractionU, bool reverseHopping, AbstractArchitecture* architecture, int memory, char* precalculationFileName, bool hoppingOnly)
+ParticleOnLatticeGenericHamiltonian::ParticleOnLatticeGenericHamiltonian(ParticleOnLattice* particles, int nbrParticles, LatticePhases *latticeGeometry, int nbrFluxQuanta, double contactInteractionU, bool reverseHopping, AbstractArchitecture* architecture, unsigned long memory, char* precalculationFileName, bool hoppingOnly)
 {
   this->Particles=particles;
   this->NbrParticles=nbrParticles;
@@ -90,7 +90,7 @@ ParticleOnLatticeGenericHamiltonian::ParticleOnLatticeGenericHamiltonian(Particl
     {
       if (memory > 0)
 	{
-	  int TmpMemory = this->FastMultiplicationMemory(memory);
+	  long TmpMemory = this->FastMultiplicationMemory(memory);
 	  PrintMemorySize(cout, TmpMemory)<< endl;
 	  if (memory > 0)
 	    this->EnableFastMultiplication();
